Working on popup and search bar
