+++ client/components/Map/Map.module.css
@@ -3,3 +3,6 @@
   width: 100vw;
   height: 100vh;
 }
+.container:focus {
+  outline: none;
+}
