+++ client/components/Map/roadMap.ts
@@ -1,0 +1,6 @@
+import Map from '@arcgis/core/Map';
+import buildingsLayer from './3DLayer';
+export default new Map({
+  basemap: 'osm',
+  layers: [buildingsLayer],
+});
