+++ client/pages/_app.tsx
@@ -1,5 +1,5 @@
 import { AppProps } from 'next/app';
-import '../styles.css';
+import '../styles.scss';
 function MyApp({ Component, pageProps }: AppProps) {
   return (
     <>
