+++ client/pages/index.tsx
@@ -1,4 +1,4 @@
-import Map from 'components/Map';
+import Map from '../components/Map/Map';
 
 const Main = () => {
   return <Map />;
