+++ client/styles.scss
@@ -349,3 +349,10 @@ template {
 [hidden] {
   display: none;
 }
+
+/* .esri-popup__main-container {
+  display: none !important;
+}
+.esri-popup__pointer-direction {
+  display: none !important;
+} */
